Update decrypt-example.rst
I noticed that the indentation of the first Python code snippet is not right on http://jug.readthedocs.org/en/latest/decrypt-example.html, and for some reason its syntax highlighting too (it's strange because the other blocks seem ok). So I simply propose to use explicit `code-block` directives. Thanks for the nice module by the way!