ENH+TST Add a side process to keep locks alive and help distinguish crashed jobs
An attempt was made on implementing this behavior with threads but due
to the GIL threads would get suspended and locks wouldn't be refreshed.

Using a separate process makes it more expensive to get/release a lock
but makes the timestamp a reliable source of information.

This new functionality is only available with file_keepalive_store which
inherits most behavior from the already existing file_store.